Feat/footer (#19)
* refactor: rename footer name field to title and update footer component styling

* feat: add privacy, cookies, and legal links to footer copyright section

* fixes build warnings

* api, footer, page, page list refactor

* got vercel building with the master environment temporarily

* swapped to development because of API restrictions

* variable api environment update

* api lint fix

---------
