+++ src/lib/contentful-api/external-page.ts
@@ -1,0 +1,6 @@
+import { SYS_FIELDS } from './constants';
+
+export const EXTERNAL_PAGE_FIELDS = `
+  ${SYS_FIELDS}
+  title
+`;
