+++ src/lib/contentful-api/index.ts
@@ -1,8 +1,14 @@
 export * from './banner-hero';
+export * from './constants';
 export * from './content-grid';
 export * from './cta-banner';
+export * from './external-page';
+export * from './footer';
+export * from './header';
 export * from './image-between';
 export * from './image';
+export * from './page';
+export * from './page-list';
 export * from './post';
 export * from './product';
 export * from './section-heading';
