+++ src/types/contentful/index.ts
@@ -1,19 +1,21 @@
 export * from './Asset';
 export * from './BannerHero';
 export * from './Button';
-export * from './CtaBanner';
 export * from './ContentGrid';
+export * from './CtaBanner';
 export * from './Footer';
 export * from './GraphQLResponse';
 export * from './Header';
 export * from './Image';
 export * from './ImageBetween';
+export * from './Modal';
 export * from './Page';
 export * from './PageList';
 export * from './Post';
 export * from './Product';
 export * from './SectionHeading';
+export * from './Service';
 export * from './Slider';
+export * from './Social';
 export * from './Solution';
 export * from './Video';
-export * from './Service';
